# api-server/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Commit-Relay Dashboard API",
    "description": "Real-time metrics and monitoring API for the commit-relay master-worker system",
    "version": "1.0.0",
    "contact": {
      "name": "Commit-Relay Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5001/api/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/tasks/{taskId}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get task details",
        "operationId": "getTask",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "in_progress", "completed", "failed"]
          },
          "priority": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "assigned_worker": {
            "type": "string"
          }
        }
      },
...
    }
  }
}